Task #34, refactor: rename generator to service
